fix matching error
I don't know why this university was unmatched before. The name is the same, the number of publications broadly aligns (for 2016-2019: 820 in Leiden ranking, 1345 in OpenAlex, but given bigger coverage in OpenAlex and stricter inclusion rules in Leiden Ranking this seems within reasonable bounds), and the location according to the map in the Leiden Ranking also fits.